Compressor pwm (#1228)
* add comp freq as pwm and delete the gpio one

* try the pwm

* update freq

* update parametric

* changed set up more

* well it went high

* it worksgit add .

* a bit of cleanup

* erd was in wrong ds

* mapped erd working

* fix up compressor speed driver

* back in business

* create a new type

* haha

* tried with out that if check

* add some comments

* changes

* ignore the type

* undo parametric change

* undo more

* oopsies

* change data source

* fix

+++ test/Tests/Application/Compressor/CompressorSpeedDriver_Test.cpp
@@ -168,9 +168,9 @@ TEST_GROUP(CompressorSpeedDriver_VariableSpeedCoolingModeIndependentAndRelayEnab
       DataModel_Write(dataModel, Erd_CompressorControllerSpeedRequest, &speed);
    }
 
-   void CompressorFrequencyShouldBe(uint8_t expected)
+   void CompressorFrequencyShouldBe(uint16_t expected)
    {
-      uint8_t actual;
+      uint16_t actual;
       DataModel_Read(dataModel, Erd_CompressorInverterDriver, &actual);
 
       CHECK_EQUAL(expected, actual);
@@ -364,9 +364,9 @@ TEST_GROUP(CompressorSpeedDriver_VariableSpeedCoolingModeDependentAndRelayDisabl
       DataModel_Write(dataModel, Erd_FreezerSetpointZone, &setpointZone);
    }
 
-   void CompressorFrequencyShouldBe(uint8_t expected)
+   void CompressorFrequencyShouldBe(uint16_t expected)
    {
-      uint8_t actual;
+      uint16_t actual;
       DataModel_Read(dataModel, Erd_CompressorInverterDriver, &actual);
 
       CHECK_EQUAL(expected, actual);
